+++ Patches/Tameable_GetHoverText_Patch.cs
@@ -14,18 +14,29 @@ namespace HoldToCommand.Patches
         static void Postfix(Tameable __instance, ref string __result)
         {
             if (!__instance.IsTamed() || !__instance.m_commandable)
+            {
                 return;
+            }
+
+            if (!Plugin.RegisterTranslations(Localization.instance))
+            {
+                return;
+            }
 
             var pet = Localization.instance.Localize("$hud_pet");
             int idx = __result.IndexOf(pet, StringComparison.Ordinal);
-            if (idx < 0) return;
+            if (idx < 0)
+            {
+                return;
+            }
 
             var useKey = Localization.instance.Localize("$KEY_Use");
             var holdUse = Localization.instance.Localize($"${Plugin.LangKeyHold}", useKey);
             var command = Localization.instance.Localize($"${Plugin.LangKeyCommand}");
-            var insert = $"  [<color=yellow><b>{holdUse}</b></color>] {command}";
+            var insert = $"[<color=yellow><b>{holdUse}</b></color>] {command}";
+            var seperator = Plugin.Configs.ShowInNewLine.Value == true ? "\n" : "  ";
 
-            __result = __result.Insert(idx + pet.Length, insert);
+            __result = __result.Insert(idx + pet.Length, seperator + insert);
         }
     }
 }