+++ Plugin.Info.cs
@@ -1,11 +1,11 @@
 using BepInEx;
 namespace HoldToCommand
 {
-	[BepInPlugin("Vonny1412.HoldToCommand", "HoldToCommand", "1.0.1")]
+	[BepInPlugin("Vonny1412.HoldToCommand", "HoldToCommand", "1.0.2")]
 	public sealed partial class Plugin
 	{
 		// generated file – do not edit
-		public const string Version = "1.0.1";
+		public const string Version = "1.0.2";
 		public const string ModName = "HoldToCommand";
 		public const string ModGuid = "Vonny1412.HoldToCommand";
 	}